Fixed the data binding of the PlayListSongSelectionEditView ListView to be an ObservableCollection<Song> filled out by the ViewModel (instead of the placeholder direct, non-updating binding to the all songs playlist)
PlayListCreationView and PlayListPlayBackView take up the same page real estate. PlayListCreationView is Visible by default, PlayListPlayBackView is Collapsed/hidden by default, these visible states are managed in MainPage event handlers

Added PlayListPlayBackView ListView...this is the ui control that appears on the right side of the page and shows the user the songs in the selected playlist. Collapsed/hidden by default

Dropped in the MediaPlayerElement for layout purposes. Doesn't work with any of the sample media files that I've tried with it